requirement.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
PyJWT==1.5.2
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
